+++ src/app/config/app.api.ts
@@ -1,0 +1,3 @@
+export const AppApi = {
+  GEOLOCATION: 'http://www.mapquestapi.com/geocoding/v1/'
+};