+++ src/app/shared/models/geocode/geocode-info-copyright.model.ts
@@ -1,0 +1,5 @@
+export class GeocodeInfoCopyright {
+  text: string;
+  imageUrl: string;
+  imageAltText: string;
+}