+++ src/app/shared/models/geocode/geocode-info.model.spec.ts
@@ -1,0 +1,7 @@
+import { GeocodeInfo } from './geocode-info.model';
+
+describe('GeocodeInfo', () => {
+  it('should create an instance', () => {
+    expect(new GeocodeInfo()).toBeTruthy();
+  });
+});