+++ src/app/shared/models/geocode/geocode-options.model.spec.ts
@@ -1,0 +1,7 @@
+import { GeocodeOptions } from './geocode-options.model';
+
+describe('GeocodeOptions', () => {
+  it('should create an instance', () => {
+    expect(new GeocodeOptions()).toBeTruthy();
+  });
+});