+++ src/app/shared/models/geocode/geocode-options.model.ts
@@ -1,0 +1,5 @@
+export class GeocodeOptions {
+  maxResults: number;
+  thumbMaps: boolean;
+  ignoreLatLngInput: boolean;
+}