+++ src/app/shared/models/geocode/geocode-result-location.model.spec.ts
@@ -1,0 +1,7 @@
+import { GeocodeResultLocation } from './geocode-result-location.model';
+
+describe('GeocodeResultLocation', () => {
+  it('should create an instance', () => {
+    expect(new GeocodeResultLocation()).toBeTruthy();
+  });
+});