+++ src/app/shared/models/geocode/geocode-result-provided-location.model.ts
@@ -1,0 +1,3 @@
+export class GeocodeResultProvidedLocation {
+  latLng: { lat: number; lng: number };
+}