+++ src/app/shared/models/geocode/geocode-result.model.ts
@@ -1,0 +1,7 @@
+import { GeocodeResultProvidedLocation } from './geocode-result-provided-location.model';
+import { GeocodeResultLocation } from './geocode-result-location.model';
+
+export class GeocodeResult {
+  providedLocation: GeocodeResultProvidedLocation;
+  locations: GeocodeResultLocation[];
+}