+++ src/app/shared/models/navigator/navigator-geolocation-position.model.ts
@@ -3,12 +3,4 @@ import { NavigatorGeolocationPositionCoords } from './navigator-geolocation-posi
 export class NavigatorGeolocationPosition {
   coords: NavigatorGeolocationPositionCoords;
   timestamp: number;
-
-  constructor(position?: NavigatorGeolocationPosition) {
-    this.coords =
-      position && position.coords
-        ? new NavigatorGeolocationPositionCoords(position.coords)
-        : new NavigatorGeolocationPositionCoords();
-    this.timestamp = position ? position.timestamp : null;
-  }
 }