+++ src/app/shared/services/countries-web.service.ts
@@ -1,8 +1,11 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { Country } from '../models/country.model';
+
+// Stubs
+import countriesStub from 'src/assets/data/stubs/stub-countries.json';
 
 @Injectable({
   providedIn: 'root'
@@ -17,6 +20,11 @@ export class CountriesWebService {
   constructor(private http: HttpClient) {}
 
   public getCountries(): Observable<Country[]> {
+    return of(countriesStub).pipe(
+      map((countries) => {
+        return countries.map((country: any) => new Country(country.alpha2Code, country.translations[this.lang]));
+      })
+    );
     return this.http.get<object[]>(`${this.api}/all?fields=alpha2Code;translations`).pipe(
       map((countries) => {
         return countries.map((country: any) => new Country(country.alpha2Code, country.translations[this.lang]));